从distributor分离出request_recorder
